Added Results for another image
For the same image, prime potrait was calculated for two dimensions
107x53 and 53x53, when prime potrait for 107x53 was calculated,
it was very fuzzy, but prime potrait for 53x53 was comparitively
a lot better, atleast we could see some features.

One of the conclusions i came to was that, it the image is a square
one (i.e, one of equal dimensions), then output prime potrait would
be a lot better compared to say a rectangular one.